+++ packages/lumx-vue/src/components/text-field/index.ts
@@ -1,0 +1,2 @@
+export { default as RawInputText, type RawInputTextProps } from './RawInputText';
+export { default as RawInputTextarea, type RawInputTextareaProps } from './RawInputTextarea';